sample management tool
added weights to output json file, fixed timestamping and masked out broken folders, deleted old messy working copies of code and unneccesary reference files, added comments explaining code.
