--- SampleManagement/folders.py
@@ -1,44 +1,0 @@
-#!/usr/bin/env python
-
-folderNames = [
-	#data
-	'DoubleMuon',
-
-	#signal
-	'CRAB_PrivateMC',
-
-	#TTJets
-	'TTJets_TuneCP5_13TeV-madgraphMLM-pythia8',
-
-	#diboson
-	'WW_TuneCP5_13TeV-pythia8',
-	'WZ_TuneCP5_13TeV-pythia8',
-	'ZZ_TuneCP5_13TeV-pythia8',
-
-	#triboson
-	'WGG_5f_TuneCP5_13TeV-amcatnlo-pythia8',
-	'WWG_TuneCP5_13TeV-amcatnlo-pythia8',
-	'WWW_4F_TuneCP5_13TeV-amcatnlo-pythia8',
-	'WWZ_TuneCP5_13TeV-amcatnlo-pythia8',
-	'WZG_TuneCP5_13TeV-amcatnlo-pythia8',
-	'WZZ_TuneCP5_13TeV-amcatnlo-pythia8',
-	'ZZZ_TuneCP5_13TeV-amcatnlo-pythia8',
-
-	#DYJetsToLL
-	'DYJetsToLL_M-10to50_TuneCP5_13TeV-madgraphMLM-pythia8',
-	'DYJetsToLL_M-50_TuneCP5_13TeV-madgraphMLM-pythia8',
-	'QCD_Pt-15to20_MuEnrichedPt5_TuneCP5_13TeV_pythia8',
-	'QCD_Pt-20to30_MuEnrichedPt5_TuneCP5_13TeV_pythia8',
-	'QCD_Pt-30to50_MuEnrichedPt5_TuneCP5_13TeV_pythia8',
-	'QCD_Pt-50to80_MuEnrichedPt5_TuneCP5_13TeV_pythia8',
-	'QCD_Pt-80to120_MuEnrichedPt5_TuneCP5_13TeV_pythia8',
-	'QCD_Pt-120to170_MuEnrichedPt5_TuneCP5_13TeV_pythia8',
-	'QCD_Pt-170to300_MuEnrichedPt5_TuneCP5_13TeV_pythia8',
-	'QCD_Pt-300to470_MuEnrichedPt5_TuneCP5_13TeV_pythia8',
-	'QCD_Pt-470to600_MuEnrichedPt5_TuneCP5_13TeV_pythia8',
-	'QCD_Pt-600to800_MuEnrichedPt5_TuneCP5_13TeV_pythia8',
-	'QCD_Pt-800to1000_MuEnrichedPt5_TuneCP5_13TeV_pythia8',
-	'QCD_Pt-1000toInf_MuEnrichedPt5_TuneCP5_13TeV_pythia8'
-
-
-]
